+++ tests/test_interact.py
@@ -11,13 +11,13 @@ from ._common import XmlRpcTestCase
 class TestInteract(XmlRpcTestCase):
     server_version = '6.1'
     startup_calls = (
-        call(ANY, 'db', ANY, verbose=ANY),
+        call(ANY, 'db', ANY, None, verbose=ANY),
         'db.server_version',
-        call(ANY, 'db', ANY, verbose=ANY),
-        call(ANY, 'common', ANY, verbose=ANY),
-        call(ANY, 'object', ANY, verbose=ANY),
-        call(ANY, 'report', ANY, verbose=ANY),
-        call(ANY, 'wizard', ANY, verbose=ANY),
+        call(ANY, 'db', ANY, None, verbose=ANY),
+        call(ANY, 'common', ANY, None, verbose=ANY),
+        call(ANY, 'object', ANY, None, verbose=ANY),
+        call(ANY, 'report', ANY, None, verbose=ANY),
+        call(ANY, 'wizard', ANY, None, verbose=ANY),
         'db.list',
     )
 
@@ -31,6 +31,7 @@ class TestInteract(XmlRpcTestCase):
         mock.patch('erppeek.Client._set_interactive', wraps=erppeek.Client._set_interactive).start()
         self.interact = mock.patch('erppeek._interact', wraps=erppeek._interact).start()
         self.infunc = mock.patch('code.InteractiveConsole.raw_input').start()
+        mock.patch('erppeek.main.__defaults__', (self.interact,)).start()
 
     def test_main(self):
         env_tuple = ('http://127.0.0.1:8069', 'database', 'usr', None)
@@ -104,6 +105,7 @@ class TestInteract(XmlRpcTestCase):
     def test_invalid_user_password(self):
         env_tuple = ('http://127.0.0.1:8069', 'database', 'usr', 'passwd')
         mock.patch('sys.argv', new=['erppeek', '--env', 'demo']).start()
+        mock.patch('os.environ', new={'LANG': 'fr_FR.UTF-8'}).start()
         mock.patch('erppeek.read_config', return_value=env_tuple).start()
         mock.patch('getpass.getpass', return_value='x').start()
         self.service.db.list.return_value = ['database']
@@ -121,11 +123,13 @@ class TestInteract(XmlRpcTestCase):
         usr17 = ('object.execute', 'database', 17, 'passwd')
         expected_calls = self.startup_calls + (
             ('common.login', 'database', 'usr', 'passwd'),
-            usr17 + ('ir.model', 'search', [('model', 'like', 'res.company')]),
+            usr17 + ('ir.model', 'search',
+                     [('model', 'like', 'res.company')]),
             usr17 + ('ir.model', 'read', 42, ('model',)),
             usr17 + ('ir.model.access', 'check', 'res.users', 'write'),
             ('common.login', 'database', 'gaspard', 'x'),
-            usr17 + ('ir.model', 'search', [('model', 'like', 'res.company')]),
+            usr17 + ('ir.model', 'search',
+                     [('model', 'like', 'res.company')]),
             usr17 + ('ir.model', 'read', 42, ('model',)),
         )
         self.assertCalls(*expected_calls)